workingfrom postman to drone api #4

+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/api/TabletRestController.java
@@ -13,7 +13,7 @@ public class TabletRestController {
     DroneLauncher droneLauncher;
 
     @PostMapping("/start/drone/{droneId}/package/{packageId}")
-    public void start(@PathVariable int droneId, @PathVariable int packageId){
+    public void start(@PathVariable Long droneId, @PathVariable Long packageId){
         droneLauncher.start(droneId, packageId);
     }
