+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/interfaces/DroneFinder.java
@@ -4,5 +4,5 @@ import com.polytech.si5.al.dronedelivery.team.g.truck.entities.Drone;
 
 public interface DroneFinder {
 
-    public Drone findDroneById(int droneId);
+    public Drone findDroneById(Long droneId);
 }
