+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/interfaces/DroneLauncher.java
@@ -2,5 +2,5 @@ package com.polytech.si5.al.dronedelivery.team.g.truck.interfaces;
 
 public interface DroneLauncher {
 
-    public void start(int droneId, int packageId);
+    public void start(Long droneId, Long packageId);
 }
